+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed rollup-memory-range.lua to input/output JSON objects
 - Changed to LuaRocks for Lua dependencies
 - Changed grpc lib version to v1.45
+- Changed dhd and rollup fields in config to optional
 
 ## [0.8.0] - 2021-12-28